--- example_for_testing/Cargo.toml
@@ -1,25 +1,0 @@
-[package]
-name = "macroquad-3d"
-version = "0.1.0"
-authors = ["DodoRare Team <support@dodorare.com>"]
-edition = "2018"
-
-[dependencies]
-crossbow = { version = "0.1.0", path = "../../" }
-log = "0.4"
-anyhow = "1.0"
-macroquad = "0.3"
-
-[package.metadata]
-app_name = "Macroquad 3D"
-target_sdk_version = 30
-version_code = "1"
-icon = "ic_launcher"
-
-android_build_targets = ["aarch64-linux-android"]
-android_assets = "assets"
-android_res = "res/android"
-
-apple_build_targets = ["aarch64-apple-ios", "x86_64-apple-ios"]
-apple_assets = "assets"
-apple_res = "res/apple"
