+++ tests/resources/android_app_bundle/minimal_unsigned.aab
