+++ tests/resources/manifest/AndroidManifest.xml
