+++ tests/resources/res/android/mipmap-hdpi/ic_launcher.png
